pyproject.toml
[project]
name = "movbits-api"
version = "0.1.0"
description = "Movbits API in Django / Multi-tenant"
authors = [
    {name = "Grant Viklund",email = "grant@movbits.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "django (>=6.0.4,<7.0)",
    "django-allauth (>=65.7.0,<66.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "gunicorn",
    "psycopg2-binary",
    "dj-database-url",
    "boto3",
    "cryptography (>=43.0.0)",
    "django-storages[s3] (>=1.14.6,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "whitenoise (>=6.9.0,<7.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "stripe (>=12.2.0,<13.0.0)",
    "django-anymail (>=13.0,<14.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "dj-rest-auth (>=7.0.0,<8.0.0)",
    "drf-spectacular[sidecar] (>=0.28.0,<0.29.0)",
    "django-vendor (>=0.7.0,<0.8.0)",
    "django-site-configs (>=0.4.0,<0.5.0)",
    "django-integrations (>=0.4.0,<0.5.0)",
    "python-json-logger (>=4.0.0,<5.0.0)",
    "google-cloud-pubsub (>=2.26.0,<3.0.0)",
    "google-cloud-bigquery (>=3.0.0,<4.0.0)",
    "django-invitations (>=2.0.0,<3.0.0)",
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.2.0"
mypy = "^1.15.0"
bandit = "^1.8.3"
isort = "^6.0.1"
pytest-django = "^4.11.1"
pytest = "^8.3.5"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
